WFLY-15996 Make Java 9+ version of FastConcurrentDirectQueue the default implementation.
