--- clustering/ee/cache/build-test-java11
@@ -1,1 +1,0 @@
-https://github.com/jboss/jboss-parent-pom#mr-jar-testing
