+++ ee-feature-pack/common/src/main/resources/modules/system/layers/base/org/wildfly/clustering/ee/cache/main/module.xml
@@ -44,6 +44,5 @@
         <module name="org.wildfly.clustering.service"/>
         <module name="org.wildfly.common"/>
         <module name="org.wildfly.security.elytron-private"/>
-        <module name="jdk.unsupported"/>
     </dependencies>
 </module>
